add core terminal exec method
